% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tintin-package.R
\docType{package}
\name{tintin-package}
\alias{tintin}
\alias{tintin-package}
\title{tintin: Tintin Palette Generator}
\description{
Palettes generated from Tintin covers. There is one palette per cover, with a total of 24 palettes of 5 colours each. Includes functions to interpolate colors in order to create more colors based on the provided palettes. The data is based on Cyr, et al. (2004) \doi{10.1503/cmaj.1041405} and Wikipedia \url{https://en.wikipedia.org/wiki/The_Adventures_of_Tintin}.
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
