% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tintin-package.R
\docType{data}
\name{tintin_head_trauma}
\alias{tintin_head_trauma}
\title{Tintin Head Trauma}
\format{
A data frame with 50 rows and 4 variables
}
\source{
Cyr, Antoine, Louis-Olivier Cyr, and Claude Cyr. "Acquired growth
hormone deficiency and hypogonadotropic hypogonadism in a subject with
repeated head trauma, or Tintin goes to the neurologist." CMAJ 171, no. 12
(2004): 1433-1434. The book release year was obtained from Wikipedia.
}
\usage{
tintin_head_trauma
}
\description{
Neurological traumas sustained by the subject (by book title)
}
\details{
\itemize{
\item{\code{year}}{ Release year of the comic, first ocurrence was used
for titles with a b&w and colour edition}
\item{\code{book_title}}{ English title of the books}
\item{\code{page}}{ Page number where the traumas happen}
\item{\code{cause_of_injury}}{ What causes the trauma (i.e., explosion)}
\item{\code{loss_of_consciousness_length}}{Loss of consciousness length
measured by the number of cartoon frames before subject returns to normal
activity.}
\item{\code{loss_of_consciousness_severity}}{Loss of consciousness severity
measured by the number of objects revolving above Tintin's head}
}
}
\examples{
tintin_head_trauma
}
\keyword{datasets}
